client/requirements.txt
click
pytest
pytest-mock
PyYAML
platformdirs
requests
